refactor(checker): relocate export* cycle ownership helpers to sibling module
Move value_variable_owned_by_current_file_not_foreign and
foreign_file_wildcard_reexports_current_file out of
state/type_analysis/core.rs into a dedicated
circular_export_ownership.rs sibling so core.rs returns under its
no-slack size ratchet (1815). Behavior-preserving: the immer TS7053
fix logic is unchanged.
